Camera And Dialogue additions/fixes
Added more functionality to the camera with functions aiding with movement and rotation.

Dialogue speed and portraits are easier to use. As long as there is one element then for the remainder of the conversation if the speed or portrait is missing/unspecified the first element will be set as the default case.
Additionally, the dialogue box image can now be changed in the dialogue inspector with the same functionality as above.