Inventory Updates, Code Cleanup, Journal Additions
Journal/Inventory:

Added side tabs for different chapters in the journal that appear as you progress. Selecting a tab displays the inventory for that chapter.

Created a proper display on the inventory page. Hovering over an item within the inventory will display its description on the adjacent page.

Added a level manager but right now the only thing it does is contain the current level number

Added comments, moved around assignments, general cleanup