Inspection Clues
Item Inspector:

Clues/observations can now be put onto objects that can be found when inspecting.

Clues will have a yellow glow to hint to the player where they are

When clicked on, player character will make a comment about the clue. When all clues are found, player will be able to add it to their inventory

Fixed lazy errors that appear if you start a level without a save file