Present Clues to NPCs
There is now an option in the dialogue builder to present clues at the end of the current dialogue:

When checked and all dialogue in the current box is exhausted, the players journal will open up where they can select a clue to present to the npc. Depending on if it's the clue the npc is looking for or not the dialogue will split into 2 paths.

Definitely has bugs but it's late so I didn't really bug test. Will prob bug test tomorrow