Changed Dialogue Writing To Not Look Like Shit
Dialogue now uses textmesh pro to allow for toggling alphas of characters. This fixes the gross affect of the text writing out a word on one line, realizing partway through that it won't fit then snapping it to the line below.

Additionally, when dialogue is automatically split when too big, it will be split from the last full word within the allowed size instead of by the allowed size so that words aren't broken in half in the dialogue